backend/requirements.txt
fastapi==0.99.1
uvicorn==0.23.2
pydantic==1.10.12
python-dotenv
langchain_openai
langchain
langchain_community
langchain_core
chromadb